@@ -12,20 +12,20 @@
   "private": true,
   "dependencies": {
     "@angular-devkit/build-angular": "^13.3.7",
-    "@angular/animations": "~13.3.0",
+    "@angular/animations": "~14.2.8",
     "@angular/cdk": "^13.3.8",
     "@angular/cli": "^13.3.7",
-    "@angular/common": "^13.0.3",
-    "@angular/compiler": "^13.0.3",
-    "@angular/compiler-cli": "^13.0.3",
-    "@angular/core": "^13.0.3",
+    "@angular/common": "^14.2.8",
+    "@angular/compiler": "^14.2.8",
+    "@angular/compiler-cli": "^14.2.8",
+    "@angular/core": "^14.2.8",
     "@angular/flex-layout": "^13.0.0-beta.38",
-    "@angular/forms": "^13.0.3",
+    "@angular/forms": "^14.2.8",
     "@angular/material": "^13.3.8",
     "@angular/material-moment-adapter": "^13.3.8",
-    "@angular/platform-browser": "^13.0.3",
-    "@angular/platform-browser-dynamic": "^13.0.3",
-    "@angular/router": "^13.0.3",
+    "@angular/platform-browser": "^14.2.8",
+    "@angular/platform-browser-dynamic": "^14.2.8",
+    "@angular/router": "^14.2.8",
     "@babel/core": "^7.14.2",
     "@fortawesome/angular-fontawesome": "^0.10.2",
     "@fortawesome/fontawesome-svg-core": "^1.2.35",
@@ -87,7 +87,7 @@
     "typescript": "^4.6.4"
   },
   "devDependencies": {
-    "@angular/language-service": "^13.0.3",
+    "@angular/language-service": "^14.2.8",
     "@ngrx/schematics": "^13.2.0",
     "@types/jasmine": "^3.6.11",
     "@types/jasminewd2": "^2.0.8",
